requirements.txt
requests==2.31.0
SQLAlchemy==1.4.51
pandas==2.1.4
psycopg2-binary==2.9.9
aiofiles==23.2.1
numpy==1.24.4
factory-boy==3.3.1
schedule==1.2.2
google-api-core
google-auth
boto3==1.36.26
dotenv
pytest